Update: update in-code docs and re-organize some functionality (#811)
