+++ tests/integrations/__init__.py
@@ -13,3 +13,7 @@
 # limitations under the License.
 
 # flake8: noqa
+
+from base_tester import *
+from config import *
+from helpers import *
