+++ tests/integrations/image_classification/configs/test_resnet50_imagenette_pruned-end_to_end.yaml
