+++ tests/integrations/object_detection/configs/test_all_pruned_quantized_short.yaml
