+++ tests/integrations/object_detection/test_object_detection.py
@@ -14,22 +14,22 @@
 
 import os
 import tempfile
-from collections import defaultdict
 
-import numpy
 import onnx
 import pandas as pd
 import pytest
 import torch
-from onnxruntime import InferenceSession
 
-from sparseml.onnx.utils import get_tensor_shape
 from tests.integrations.base_tester import (
     BaseIntegrationManager,
     BaseIntegrationTester,
     skip_inactive_stage,
 )
-from tests.integrations.helpers import get_configs_with_cadence
+from tests.integrations.helpers import (
+    get_configs_with_cadence,
+    test_model_inputs_outputs,
+    test_model_op_counts,
+)
 from tests.integrations.object_detection.args import Yolov5ExportArgs, Yolov5TrainArgs
 from yolov5.export import load_checkpoint
 
@@ -159,7 +159,7 @@ class TestObjectDetection(BaseIntegrationTester):
             type_="val", weights=target_model_path, device=torch.device("cpu")
         )
 
-        _test_model_op_counts(export_model_path, target_model_path)
+        test_model_op_counts(export_model_path, target_model_path)
 
         compare_outputs = export_args.test_args.get("compare_outputs", True)
         if isinstance(compare_outputs, str) and (
@@ -167,71 +167,4 @@ class TestObjectDetection(BaseIntegrationTester):
         ):
             compare_outputs = False
         if compare_outputs:
-            _test_model_inputs_outputs(export_model_path, target_model_path)
-
-
-_TEST_OPS = {
-    "MatMul",
-    "Gemm",
-    "Conv",
-    "MatMulInteger",
-    "ConvInteger",
-    "QLinearMatMul",
-    "QLinearConv",
-}
-
-
-def _test_model_op_counts(model_path_a, model_path_b):
-
-    model_a = onnx.load(model_path_a)
-    model_b = onnx.load(model_path_b)
-
-    def _get_model_op_counts(model):
-        op_counts = defaultdict(int)
-        for node in model.graph.node:
-            if node.op_type in _TEST_OPS:
-                op_counts[node.op_type] += 1
-        return op_counts
-
-    op_counts_a = _get_model_op_counts(model_a)
-    op_counts_b = _get_model_op_counts(model_b)
-
-    assert len(op_counts_a) > 0
-    assert len(op_counts_a) == len(op_counts_b)
-
-    for op, count_a in op_counts_a.items():
-        assert op in op_counts_b
-        assert count_a == op_counts_b[op]
-
-
-def _test_model_inputs_outputs(model_path_a, model_path_b):
-    # compare export and target graphs and build fake data
-    model_a = onnx.load(model_path_a)
-    model_b = onnx.load(model_path_b)
-    assert len(model_a.graph.input) == len(model_b.graph.input)
-    assert len(model_a.graph.output) == len(model_b.graph.output)
-
-    sample_input = {}
-    output_names = []
-
-    for input_a, input_b in zip(model_a.graph.input, model_b.graph.input):
-        assert input_a.name == input_b.name
-        input_a_shape = get_tensor_shape(input_a)
-        assert input_a_shape == get_tensor_shape(input_b)
-        d_type = (
-            numpy.uint8 if input_a.type.tensor_type.elem_type == 2 else numpy.float32
-        )
-        sample_input[input_a.name] = numpy.random.randn(*input_a_shape).astype(d_type)
-
-    for output_a, output_b in zip(model_a.graph.output, model_b.graph.output):
-        assert output_a.name == output_b.name
-        assert get_tensor_shape(output_a) == get_tensor_shape(output_b)
-        output_names.append(output_a.name)
-
-    # run sample forward and test absolute max diff
-    ort_sess_a = InferenceSession(model_path_a)
-    ort_sess_b = InferenceSession(model_path_b)
-    forward_output_a = ort_sess_a.run(output_names, sample_input)
-    forward_output_b = ort_sess_b.run(output_names, sample_input)
-    for out_a, out_b in zip(forward_output_a, forward_output_b):
-        assert numpy.max(numpy.abs(out_a - out_b)) <= 1e-4
+            test_model_inputs_outputs(export_model_path, target_model_path)
