+++ tests/integrations/transformers/configs/test_example_transformers.yaml
