+++ tests/integrations/transformers/test_transformers.py
@@ -18,17 +18,13 @@ import json
 import math
 import os
 import tempfile
-from collections import defaultdict
 
-import numpy
 import onnx
 import pytest
 import torch
-from onnxruntime import InferenceSession
 from transformers import AutoConfig, AutoTokenizer
 from transformers.tokenization_utils_base import PaddingStrategy
 
-from sparseml.onnx.utils import get_tensor_shape
 from sparseml.pytorch.optim.manager import ScheduledModifierManager
 from sparseml.transformers.export import load_task_model
 from sparseml.transformers.utils import SparseAutoModel
@@ -37,8 +33,12 @@ from tests.integrations.base_tester import (
     BaseIntegrationTester,
     skip_inactive_stage,
 )
-from tests.integrations.helpers import get_configs_with_cadence
-from tests.integrations.transformers.transformers_args import (
+from tests.integrations.helpers import (
+    get_configs_with_cadence,
+    test_model_inputs_outputs,
+    test_model_op_counts,
+)
+from tests.integrations.transformers.args import (
     MaskedLanguageModellingArgs,
     QuestionAnsweringArgs,
     TextClassificationArgs,
@@ -165,7 +165,7 @@ class TestTransformers(BaseIntegrationTester):
             os.path.dirname(export_args.run_args.model_path),
             export_args.run_args.onnx_file_name,
         )
-        _test_model_op_counts(export_model_path, target_model_path)
+        test_model_op_counts(export_model_path, target_model_path)
 
         compare_outputs = export_args.test_args.get("compare_outputs", True)
         if isinstance(compare_outputs, str) and (
@@ -173,8 +173,12 @@ class TestTransformers(BaseIntegrationTester):
         ):
             compare_outputs = False
         if compare_outputs:
-            _test_model_inputs_outputs(
-                export_model_path, target_model_path, manager.configs["export"]
+            test_model_inputs_outputs(
+                export_model_path,
+                target_model_path,
+                _create_bert_input,
+                model_path=export_args.run_args.model_path,
+                task=manager.task,
             )
 
 
@@ -186,66 +190,6 @@ def _load_model_on_task(model_name_or_path, model_type, task, **model_kwargs):
         "token_classification": SparseAutoModel.token_classification_from_pretrained,
     }
     return load_funcs[task](model_name_or_path, model_type=model_type, **model_kwargs)
-
-
-_TEST_OPS = {
-    "MatMul",
-    "Gemm",
-    "Conv",
-    "MatMulInteger",
-    "ConvInteger",
-    "QLinearMatMul",
-    "QLinearConv",
-}
-
-
-def _test_model_op_counts(model_path_a, model_path_b):
-
-    model_a = onnx.load(model_path_a)
-    model_b = onnx.load(model_path_b)
-
-    def _get_model_op_counts(model):
-        op_counts = defaultdict(int)
-        for node in model.graph.node:
-            if node.op_type in _TEST_OPS:
-                op_counts[node.op_type] += 1
-        return op_counts
-
-    op_counts_a = _get_model_op_counts(model_a)
-    op_counts_b = _get_model_op_counts(model_b)
-
-    assert len(op_counts_a) > 0
-    assert len(op_counts_a) == len(op_counts_b)
-
-    for op, count_a in op_counts_a.items():
-        assert op in op_counts_b
-        assert count_a == op_counts_b[op]
-
-
-def _test_model_inputs_outputs(model_path_a, model_path_b, config):
-    # compare export and target graphs and build fake data
-    model_a = onnx.load(model_path_a)
-    model_b = onnx.load(model_path_b)
-    assert len(model_a.graph.input) == len(model_b.graph.input)
-    assert len(model_a.graph.output) == len(model_b.graph.output)
-
-    sample_input = {}
-    output_names = []
-
-    sample_input = _create_bert_input(config.run_args.model_path, config.run_args.task)
-
-    for output_a, output_b in zip(model_a.graph.output, model_b.graph.output):
-        assert output_a.name == output_b.name
-        assert get_tensor_shape(output_a) == get_tensor_shape(output_b)
-        output_names.append(output_a.name)
-
-    # run sample forward and test absolute max diff
-    ort_sess_a = InferenceSession(model_path_a)
-    ort_sess_b = InferenceSession(model_path_b)
-    forward_output_a = ort_sess_a.run(output_names, sample_input)
-    forward_output_b = ort_sess_b.run(output_names, sample_input)
-    for out_a, out_b in zip(forward_output_a, forward_output_b):
-        assert numpy.max(numpy.abs(out_a - out_b)) <= 1e-4
 
 
 def _create_bert_input(model_path, task):
